requirements.txt
Flask>=3.0,<4.0
python-dotenv>=1.0
bcrypt>=4.0
passlib[bcrypt]>=1.7
pdfplumber>=0.11
PyPDF2>=3.0
